$page now supports jumping to a page
The $page command now supports jumping to a page with $page[pagenum] [article title]. The $page command goes to the first page if there is no pagenum provided. If the pagenum provided is out of bounds, it snaps the page to the first or last one based on which one is closer.